Create confirmation model
This allows us to link the record we wish to confirm with a
`confirmation` record as a proxy. This ensures we can deleted the
confirmation so that it cannot be used again.

Additionally, we just a [polymorphic][] relationship for better
flexibility, should we want to confirm additional records in the future.

[polymorphic]: https://guides.rubyonrails.org/association_basics.html#polymorphic-associations